# backend/openapi/specs/data-cleaning.yaml
openapi: "3.0.1"
info:
  title: "API Documentation"
  version: "1.0"
paths:
...
  /cleaning/tasks/{taskId}/stop:
    post:
      summary: "cleaningTasksStop"
      description: "cleaningTasksStop"
      parameters:
      - name: "taskId"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties: {}
                    description: "数据"
  /cleaning/tasks/{taskId}/execute:
    post:
      summary: "cleaningTasksStart"
      description: "cleaningTasksStart"
      parameters:
      - name: "taskId"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties: {}
                    description: "数据"
  /cleaning/tasks/{taskId}:
    get:
      summary: "cleaningTasksTaskIdGet"
      description: "cleaningTasksTaskIdGet"
      parameters:
      - name: "taskId"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "string"
                        description: ""
                      name:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      srcDatasetId:
                        type: "string"
                        description: ""
                      srcDatasetName:
                        type: "string"
                        description: ""
                      destDatasetId:
                        type: "string"
                        description: ""
                      destDatasetName:
                        type: "string"
                        description: ""
                      beforeSize:
                        type: "integer"
                        description: ""
                        format: "int64"
                      afterSize:
                        type: "integer"
                        description: ""
                        format: "int64"
                      fileCount:
                        type: "integer"
                        description: ""
                      status:
                        type: "string"
                        description: "{PENDING=PENDING, RUNNING=RUNNING, COMPLETED=COMPLETED,\
                          \ STOPPED=STOPPED, FAILED=FAILED}"
                        enum:
                        - "PENDING"
                        - "RUNNING"
                        - "COMPLETED"
                        - "STOPPED"
                        - "FAILED"
                      templateId:
                        type: "string"
                        description: ""
                      instance:
                        type: "array"
                        description: ""
                        items:
                          type: "object"
                          properties:
                            id:
                              type: "string"
                              description: ""
                            name:
                              type: "string"
                              description: ""
                            description:
                              type: "string"
                              description: ""
                            version:
                              type: "string"
                              description: ""
                            inputs:
                              type: "string"
                              description: ""
                            outputs:
                              type: "string"
                              description: ""
                            runtime:
                              type: "string"
                              description: ""
                            settings:
                              type: "string"
                              description: ""
                            isStar:
                              type: "boolean"
                              description: ""
                            createdAt:
                              type: "string"
                              description: ""
                            updatedAt:
                              type: "string"
                              description: ""
                          description: "OperatorResponse"
                      progress:
                        type: "object"
                        properties:
                          process:
                            type: "number"
                            description: ""
                          totalFileNum:
                            type: "integer"
                            description: ""
                          finishedFileNum:
                            type: "integer"
                            description: ""
                        description: ""
                      createdAt:
                        type: "string"
                        description: ""
                      startedAt:
                        type: "string"
                        description: ""
                      finishedAt:
                        type: "string"
                        description: ""
                    description: "数据"
    delete:
      summary: "cleaningTasksTaskIdDelete"
      description: "cleaningTasksTaskIdDelete"
      parameters:
      - name: "taskId"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "状态码"
                  message:
                    type: "string"
                    description: "消息"
                  data:
                    type: "object"
                    properties: {}
                    description: "数据"
...